Split parser expression keyword guards
